Add TLS support to the broker.  Don't use commandant.

+++ python/brokerlib.py
@@ -17,12 +17,6 @@
 # under the License.
 #
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-from __future__ import with_statement
-
 import collections as _collections
 import os as _os
 import proton as _proton
@@ -35,17 +29,26 @@ import sys as _sys
 import time as _time
 import tempfile as _tempfile
 
-class Broker(object):
-    def __init__(self, host, port, id=None, user=None, password=None, ready_file=None):
+class Broker:
+    def __init__(self, host, port, id=None, ready_file=None,
+                 user=None, password=None,
+                 cert=None, key=None, trust_store=None,
+                 quiet=False, verbose=False, init_only=False):
         self.host = host
         self.port = port
         self.id = id
         self.user = user
         self.password = password
         self.ready_file = ready_file
+        self.cert = cert
+        self.key = key
+        self.trust_store = trust_store
+        self.quiet = quiet
+        self.verbose = verbose
+        self.init_only = init_only
 
         if self.id is None:
-            self.id = "broker-{0}".format(_uuid.uuid4())
+            self.id = "broker-{0}".format(str(_uuid.uuid4())[:8])
 
         self.container = _reactor.Container(_Handler(self))
         self.container.container_id = self.id # XXX Obnoxious
@@ -59,8 +62,22 @@ class Broker(object):
 
             self._init_sasl_config()
 
+        if self.cert is not None:
+            if self.key is None:
+                self.fail("Both the cert and key files must be provided")
+
+            if not _os.path.isfile(self.key):
+                self.fail("Private key file {0} does not exist", self.key)
+
+            if not _os.path.isfile(self.cert):
+                self.fail("Certificate file {0} does not exist", self.cert)
+
+            if self.trust_store and not _os.path.isfile(self.trust_store):
+                self.fail("Trust store file {0} does not exist", self.trust_store)
+
     def _init_sasl_config(self):
         self._config_dir = _tempfile.mkdtemp(prefix="brokerlib-", suffix="")
+
         config_file = _os.path.join(self._config_dir, "proton-server.conf")
         sasldb_file = _os.path.join(self._config_dir, "users.sasldb")
 
@@ -91,20 +108,31 @@ class Broker(object):
         pass
 
     def error(self, message, *args):
-        _sys.stderr.write("{0}\n".format(message.format(*args)))
-        _sys.stderr.flush()
+        self.log(message, *args)
 
     def fail(self, message, *args):
         self.error(message, *args)
         _sys.exit(1)
 
+    def log(self, message, *args):
+        message = message[0].upper() + message[1:]
+        message = message.format(*args)
+        message = "{0}: {1}".format(self.id, message)
+
+        _sys.stderr.write("{0}\n".format(message))
+        _sys.stderr.flush()
+
     def run(self):
-        self.container.run()
+        try:
+            if self.init_only:
+                return
 
-        if _os.path.exists(self._config_dir):
-            _shutil.rmtree(self.dir, ignore_errors=True)
+            self.container.run()
+        finally:
+            if self._config_dir and _os.path.exists(self._config_dir):
+                _shutil.rmtree(self.dir, ignore_errors=True)
 
-class _Queue(object):
+class _Queue:
     def __init__(self, broker, address):
         self.broker = broker
         self.address = address
@@ -175,6 +203,17 @@ class _Handler(_handlers.MessagingHandler):
 
     def on_start(self, event):
         interface = "{0}:{1}".format(self.broker.host, self.broker.port)
+
+        if self.broker.cert is not None:
+            server_ssl_domain = event.container.ssl.server
+            server_ssl_domain.set_credentials(self.broker.cert, self.broker.key, None)
+
+            if self.broker.trust_store:
+                server_ssl_domain.set_trusted_ca_db(self.broker.trust_store)
+                server_ssl_domain.set_peer_authentication(_proton.SSLDomain.VERIFY_PEER_NAME,
+                                                          self.broker.trust_store)
+            else:
+                server_ssl_domain.set_peer_authentication(_proton.SSLDomain.ANONYMOUS_PEER)
 
         self.acceptor = event.container.listen(interface)
 
@@ -335,13 +374,13 @@ def wait_for_broker(ready_file, timeout=30):
         else:
             print("Still waiting for the broker")
 
-if __name__ == "__main__":
+def main():
     import argparse
 
-    parser = argparse.ArgumentParser()
+    parser = argparse.ArgumentParser(description="An AMQP message broker for testing")
 
-    parser.add_argument("--host", metavar="HOST", default="127.0.0.1",
-                        help="Listen for connections on HOST (default 127.0.0.1)")
+    parser.add_argument("--host", metavar="HOST", default="localhost",
+                        help="Listen for connections on HOST (default localhost)")
     parser.add_argument("--port", metavar="PORT", default=5672, type=int,
                         help="Listen for connections on PORT (default 5672)")
     parser.add_argument("--id", metavar="ID",
@@ -352,22 +391,52 @@ if __name__ == "__main__":
     #                   help="Require USER")
     # parser.add_argument("--password", metavar="SECRET",
     #                   help="Require SECRET")
+    parser.add_argument("--cert", metavar="FILE",
+                        help="The TLS certificate file")
+    parser.add_argument("--key", metavar="FILE",
+                        help="The TLS private key file")
+    parser.add_argument("--trust-store", metavar="FILE",
+                        help="The file containing trusted client certificates.  "
+                        "If set, the server verifies client certificates.")
+    parser.add_argument("--quiet", action="store_true",
+                        help="Print no logging to the console")
+    parser.add_argument("--verbose", action="store_true",
+                        help="Print detailed logging to the console")
+    parser.add_argument("--init-only", action="store_true",
+                      help=argparse.SUPPRESS)
 
     args = parser.parse_args()
 
-    def _print(message, *args):
-        message = message.format(*args)
-        _sys.stderr.write("{0}\n".format(message))
-        _sys.stderr.flush()
-
     class _Broker(Broker):
-        def info(self, message, *args): _print(message, *args)
-        def notice(self, message, *args): _print(message, *args)
-        def warn(self, message, *args): _print(message, *args)
+        def debug(self, message, *args):
+            if self.verbose:
+                self.log(message, *args)
 
-    broker = _Broker(args.host, args.port, id=args.id, ready_file=args.ready_file)
+        def info(self, message, *args):
+            if self.verbose:
+                self.log(message, *args)
+
+        def notice(self, message, *args):
+            if not self.quiet:
+                self.log(message, *args)
+
+        def warn(self, message, *args):
+            message = "Warning! {0}".format(message)
+            self.log(message, *args)
+
+        def error(self, message, *args):
+            message = "Error! {0}".format(message)
+            self.log(message, *args)
+
+    broker = _Broker(args.host, args.port, id=args.id, ready_file=args.ready_file,
+                     # user=args.user, password=args.password,
+                     cert=args.cert, key=args.key, trust_store=args.trust_store,
+                     quiet=args.quiet, verbose=args.verbose, init_only=args.init_only)
 
     try:
         broker.run()
     except KeyboardInterrupt:
         pass
+
+if __name__ == "__main__":
+    main()
